Correções menores

+++ src/main/java/com/kuuhaku/command/commands/discord/hero/ExpeditionsCommand.java
@@ -77,7 +77,7 @@ public class ExpeditionsCommand implements Executable {
 
 			for (Map.Entry<String, Object> entry : e.getRewards().entrySet()) {
 				Reward rew = Reward.valueOf(entry.getKey());
-				int val = (int) entry.getValue();
+				int val = (int) (double) entry.getValue();
 
 				eb.addField(rew.toString(),
 						switch (rew) {
